+++ rc/pksrc.h
@@ -38,6 +38,7 @@
 #define DLGCOMPAREFILES                 559
 #define IDC_GENERIC2                    1001
 #define IDC_CHECK_ENCRYPT               1003
+#define IDC_DOCTYPES                    1004
 #define IDC_CUSTOM1                     1006
 #define IDD_PREVIEW_FRAME               1006
 #define IDC_STRING1                     1007
@@ -57,7 +58,7 @@
 // 
 #ifdef APSTUDIO_INVOKED
 #ifndef APSTUDIO_READONLY_SYMBOLS
-#define _APS_NEXT_RESOURCE_VALUE        218
+#define _APS_NEXT_RESOURCE_VALUE        219
 #define _APS_NEXT_COMMAND_VALUE         40037
 #define _APS_NEXT_CONTROL_VALUE         1018
 #define _APS_NEXT_SYMED_VALUE           101